requirements.txt
pandas
numpy
scikit-learn
mlflow
python-dotenv
sqlalchemy
psycopg2-binary
matplotlib
streamlit
joblib